+++ Source/NSDate.m
@@ -1304,7 +1304,7 @@ otherTime(NSDate* other)
   return self;
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
   return (unsigned)[self timeIntervalSinceReferenceDate];
 }
@@ -1460,7 +1460,7 @@ otherTime(NSDate* other)
   return self;
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
   return (unsigned)_seconds_since_ref;
 }
